refactor: move non-test CLI scripts from tests/ into scripts/ (#5)
Eight standalone argparse CLIs living in tests/ were never pytest
tests (none match test_*.py); they blurred the seam between test
contracts and research/benchmark tooling. Moved:

  tests/run_has_text_2_detector_threshold_sweep.py     -> scripts/has_text_2/run_detector_threshold_sweep.py
  tests/analyze_has_text_2_detector_threshold_sweep.py -> scripts/has_text_2/analyze_detector_threshold_sweep.py
  tests/build_has_text_2_replay_matrix.py              -> scripts/has_text_2/build_replay_matrix.py
  tests/evaluate_has_text_2_replay_policy.py           -> scripts/has_text_2/evaluate_replay_policy.py
  tests/harvest_has_text_2_detections.py               -> scripts/has_text_2/harvest_detections.py
  tests/harvest_has_text_2_pipeline_bboxes.py           -> scripts/has_text_2/harvest_pipeline_bboxes.py
  tests/run_low_confidence_rescue.py                   -> scripts/has_text_2/run_low_confidence_rescue.py
  tests/generate_text_watermark_fixtures.py            -> scripts/fixtures/generate_text_watermark_fixtures.py

Follow-on fixes required by the move (not just git mv):
- Sibling bare imports of harvest_has_text_2_pipeline_bboxes updated
  to the new sibling filename harvest_pipeline_bboxes across all 5
  scripts that import it, including the try/except pytest-import
  fallback branches (now scripts.has_text_2.harvest_pipeline_bboxes
  instead of tests.harvest_has_text_2_pipeline_bboxes).
- run_detector_threshold_sweep.py self-invokes via subprocess with
  cwd=Path(__file__).resolve().parents[N] to land on the repo root;
  N was 1 for tests/<file>.py and had to become 2 for the new
  scripts/has_text_2/<file>.py depth (two call sites).
- generate_text_watermark_fixtures.py's ROOT = Path(__file__)...
  had the same one-level-deeper problem; .parent.parent became
  .parent.parent.parent.
- tests/test_replay_matrix.py (a real test, stays in tests/) imports
  from two of the moved modules; updated from tests.X to
  scripts.has_text_2.X.
- scripts/zero_sweep_watchdog.ps1's hardcoded paths and process-match
  string updated to the new locations (the sweep it drives already
  completed; SWEEP_COMPLETE.marker present, no scheduled task
  currently registered, so this was safe to edit without disrupting
  a live run).

pyproject.toml required no changes: testpaths was already tests-only
(these files were never collected as tests either way) and
[tool.basedpyright].include already lists both scripts and tests.
No README/docs/CI references existed to update.

Verification:
- uv run ruff check untextre tests scripts -> All checks passed!
- uv run pytest -x -q --collect-only -> 399 tests collected, 0 errors
- uv run pytest tests/test_replay_matrix.py -m "not slow" -> 5 passed
  (proves the new scripts.has_text_2.* import path resolves)
- uv run pytest -m "not slow" -> 372 passed, 2 skipped, 1 failed
  (test_cli.py::TestMainForceBbox::test_force_bbox_valid_parse hit
  the same cv2.error OutOfMemoryError machine flake seen in prior
  sessions; reran alone -> passed. Unrelated to this diff, touches
  no file changed here.)
- uv run basedpyright on moved files shows only pre-existing
  Optional/tuple-size errors of the same class issue #2 is scoped to
  fix; no new import-resolution errors from the move.
