+++ scripts/has_text_2/harvest_detections.py
