+++ scripts/has_text_2/harvest_pipeline_bboxes.py
